# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rabbit JetStream Management API
  version: 1.0.0
  description: Stable v1 control-plane and operations contract. Resource GET/HEAD requires operator or auditor authentication by default (401 missing/invalid credentials, 403 insufficient role, 404 read_api_disabled when authentication is unconfigured). Authenticated resource operations accept X-RJS-Tenant; it is required when an identity belongs to multiple tenants, defaults to the sole membership otherwise, and is checked before any tenant backend access. Only explicit loopback-bound RJS_LOCAL_DEMO permits anonymous resource reads; session, audit and writes remain authenticated. Probes, metrics, static UI and published contracts remain public and require network-level exposure controls.
servers:
  - url: /
security: [{bearerAuth: []}]
paths:
...
  /api/v1/streams:
    get:
      operationId: listStreams
      description: Filter the complete successful enumeration by name, sort, then paginate. Total is the filtered count. Unknown/repeated/malformed query parameters return 400. Not an atomic snapshot or an indexed-query performance guarantee.
      parameters: [{$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/NameQuery'}, {$ref: '#/components/parameters/NameSort'}, {$ref: '#/components/parameters/SortOrder'}]
      responses: {'200': {$ref: '#/components/responses/JSON'}, '400': {$ref: '#/components/responses/Error'}, '503': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
...
    NameQuery: {name: q, in: query, description: Trimmed case-insensitive literal name substring; maximum 256 normalized UTF-8 bytes., schema: {type: string}}
    NameSort: {name: sort, in: query, schema: {type: string, enum: [name], default: name}}
    SortOrder: {name: order, in: query, schema: {type: string, enum: [asc, desc], default: asc}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
...
  responses:
    JSON: {description: Successful JSON response, content: {application/json: {schema: {type: object}}}}
    Error: {description: Stable error envelope, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            mutation:
              $ref: '#/components/schemas/MutationAttemptEvidence'
            issues_version:
              type: string
              const: rjs.queue-validation.v1
              description: Present with issues for semantic Queue validation failures only; decoding errors retain text without guessed paths.
            issues:
              type: array
              description: Paths refer to the submitted document before canonical sorting. Group-level constraints can point to a parent object.
              items:
                type: object
                required: [path, code, message]
                properties:
                  path: {type: string, description: JSON Pointer; may identify a missing required field.}
                  code: {type: string, description: Stable semantic rule identifier.}
                  message: {type: string}
    MutationAttemptEvidence:
      type: object
      required: [schemaVersion, scope, phase, resourceEffects]
      description: Optional PUT/DELETE failure evidence for this receiving handler attempt only. Covers Queue/Stream/Consumer and declaration effects; excludes audit/lock metadata and earlier transport replays. Not an idempotency key, rollback assertion or durable original-request outcome. Absence conveys no phase information. Existing status/code/message semantics are unchanged.
      properties:
        schemaVersion: {type: string, const: rjs.mutation-evidence.v1}
        scope: {type: string, const: receiving-attempt}
        phase: {type: string, enum: [audit_intent, backend, audit_outcome]}
        resourceEffects:
          type: string
          enum: [none, possible]
          description: none only for audit_intent failure before backend invocation; possible for backend or audit_outcome failure, including conflicts. Never authorizes automatic retry.
        intentId:
          type: string
          pattern: '^[0-9a-f]{32}$'
          description: Audit intent correlation when generated. Persistence is not guaranteed, including after a lost publish acknowledgment.
      oneOf:
        - properties:
            phase: {const: audit_intent}
            resourceEffects: {const: none}
        - properties:
            phase: {enum: [backend, audit_outcome]}
            resourceEffects: {const: possible}
...